% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_csts_n.R
\docType{data}
\name{comments_csts_n}
\alias{comments_csts_n}
\title{Comment-level cross-sectional time-series data by notifier}
\format{
A data frame with 6 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{\code{numeric}. The year the member state or institution
submitted the comment.}
\item{notification_by_id}{\code{numeric}. An ID number that uniquely
identifies the member state that notified the proposed technical regulation
to the Commission.}
\item{notification_by}{\code{string}. The name of the member state that
notified the proposed technical regulation to the Commission.}
\item{notification_by_code}{\code{string}. A multi-letter code assigned by
the Commission that uniquely identifies the member state that notified the
proposed technical regulation to the Commission.}
\item{count_comments}{\code{numeric}. A count of the number of comments
submitted by the third-party member state or institution.}
}
}
\usage{
comments_csts_n
}
\description{
This dataset includes aggregated data on the number of comments submitted
against each notifying member state per year (cross-sectional time-series
data). There is one observation per notifying member state per year
(1988-2020), excluding state-years where the state was not a member of the
Single Market.
}
\keyword{datasets}
