% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_net.R
\docType{data}
\name{comments_net}
\alias{comments_net}
\title{Comment-level network data}
\format{
A data frame with 7 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{\code{numeric}. The year the member state or institution
submitted the comment.}
\item{from_node_id}{\code{numeric}. An ID number that uniquely identifies
each node (a member state or institution) that creates a link.}
\item{from_node}{\code{string}. The name of the member state or institution
that submitted the opinion to the Commission.}
\item{to_node_id}{\code{numeric}. An ID number that uniquely identifies
each node (a member state) that receives a link.}
\item{to_node}{\code{string}. The name of the member state that notified
the technical regulation to the Commission.}
\item{edge_weight}{\code{numeric}. The weight of the edge, which is the
number of comments filed against the member state the notified the proposed
technical regulation.}
}
}
\usage{
comments_net
}
\description{
This dataset includes network data on comments. Network data is similar to
directed dyad-year data except that it only includes directed dyad-years
with at least one comment. For every year, there is one node per member
state or institution that can submit a comment or notify a proposed
technical regulation (only member states). Links can only exist between a
third-party member state or institution that can submit a comment and a
member state that can notify a technical regulation. There is a link
between a third-party member state or institution and a notifying member
state if and only if the former submitted at least one comment with respect
to a technical regulation notified by the latter. The weight of the edge is
the number of comments submitted. There is one observation per third-party
member state or institution per notifying member state per year
(2002-2020), excluding directed dyad-years where a state was not a member
of the Single Market, but only if count of comments is positive.
}
\keyword{datasets}
