% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_csts.R
\docType{data}
\name{notifications_csts}
\alias{notifications_csts}
\title{CSTS data on notifications}
\format{
A data frame with 6 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the member state notified the proposed
technical regulation to the Commission.}
\item{notification_by_id}{Numeric. An ID number that uniquely identifies
the member state that notified the proposed technical regulation to the
Commission.}
\item{notification_by}{String. The name of the member state that notified
the proposed technical regulation to the Commission.}
\item{notification_by_code}{String. A multi-letter code assigned by the
Commission that uniquely identifies the member state that notified the
proposed technical regulation to the Commission.}
\item{count_notifications}{Numeric. A count of the number of notifications
made by the member state to the Commission.}
}
}
\usage{
notifications_csts
}
\description{
This dataset includes aggregated data on the number of notified technical
regulations per member state per year (cross-sectional time-series data).
There is one observation per notifying member state per year (2002-2020),
excluding state-years where the state was not a member of the Single
Market.
}
\keyword{datasets}
