% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opinions_ts.R
\docType{data}
\name{opinions_ts}
\alias{opinions_ts}
\title{Opinion-level time-series data}
\format{
A data frame with 3 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{\code{numeric}. The year the member state or institution
submitted the detailed opinion.}
\item{count_opinions}{\code{numeric}. A count of the number of opinions
submitted by the third-party member state or institution.}
}
}
\usage{
opinions_ts
}
\description{
This dataset includes aggregated data on the total number of detailed
opinions submitted per year (time-series data). There is one observation
per year (2002-2020).
}
\keyword{datasets}
