% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_ts.R
\docType{data}
\name{comments_ts}
\alias{comments_ts}
\title{TS data on comments}
\format{
A data frame with 3 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the member state or institution submitted the
comment.}
\item{count_comments}{Numeric. A count of the number of comments submitted
by the third-party member state or institution.}
}
}
\usage{
comments_ts
}
\description{
This dataset includes aggregated data on the total number of comments
submitted per year (time-series data). There is one observation per year
(2002-2020).
}
\keyword{datasets}
