% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications_ts.R
\docType{data}
\name{notifications_ts}
\alias{notifications_ts}
\title{TS data on notifictions}
\format{
A data frame with 3 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the member state notified the proposed
technical regulation to the Commission.}
\item{notification_by_id}{Numeric. An ID number that uniquely identifies
the member state that notified the proposed technical regulation to the
Commission.}
}
}
\usage{
notifications_ts
}
\description{
This dataset includes aggregated data on the number of notified technical
regulations per year (time-series data). There is one observation per year
(1988-2020).
}
\keyword{datasets}
